Allow oauth scopes 'ejabberd:admin' and 'ejabberd:user'
Those scopes group a set of commands that the user might want to
give access to a third party api. The pervious way (where each command
name is an implicit scope that only givess access to that command) is
still supported.
This should make it simpler for users to use OAuth facilities, as
doesn't require to list one by one all the commands, while fine-grained
delegation is still possible.